for_array をファイル分割
